Add exceptions and networking from qgis_plugin_tools and minor refactoring
