+++ Qaava/definitions/constants.py
@@ -1,16 +1,13 @@
-IDENTIFIER = "Qaava"
-QAAVA = "/Qaava"
+from ..qgis_plugin_tools.tools.resources import setting_key
 
 # Urls
 QAAVA_GITHUB_URL = "https://raw.githubusercontent.com/GispoCoding/qaava"
 
 DETAILED_PLAN_DATA_MODEL_URL = f"{QAAVA_GITHUB_URL}/master/asemakaavan-tietomalli/tietomalli_luonnos.sql"
 
-ENCODING = "utf-8"
-
 # Database
 PG_CONNECTIONS = "PostgreSQL/connections"
-QAAVA_DB_NAME = f"{QAAVA}/postgresqlConnectionName"
+QAAVA_DB_NAME = setting_key("postgresqlConnectionName")
 
 QGS_SETTINGS_PSYCOPG2_PARAM_MAP = {
     'database': 'dbname',
