+++ Qaava/test/conftest.py
@@ -14,7 +14,7 @@ from qgis.core import QgsProject
 
 from ..qgis_plugin_tools.testing.utilities import get_qgis_app
 from ..model.land_use_plan import LandUsePlanEnum
-from ..utils.constants import PG_CONNECTIONS
+from ..definitions.constants import PG_CONNECTIONS
 
 QGIS_APP, CANVAS, IFACE, PARENT = get_qgis_app()
 QGIS_INSTANCE = QgsProject.instance()
